Started playing with session/cookies
Working on implementing the features for issue #23.

It's now possible to start a XenForo session by using the "login" action
and then setting a cookie on the client's browser with the returned
cookie data, this will automaticlly log the user in when they visit the
XenForo forum.

I'll come up with an example later.