Added new action: upgradeUser
For feature #26
This action enables upgrading users by using the upgradeUser parameter.
The API will return the upgrade record and if the upgrade already exists
or if it's a new upgrade.
The upgrade will get extended by the specified default upgrade time if
the upgrade already exists for the user and end_time is not specified.