Workaround for creating a session for the client
For feature request #23
In addtion, I've removed the username/password/ip_address arguments for
the "login" action and replaced this with a "login_hash" argument, which
has to contain an encrypted (with a salt) string.
I enabled the "login_hash" argument for the "authenticate" action as
well, so it's possible to authenticate users without sending the
password in plain-text.