Fixed authentication issues with hashes
The user hash used to be seperated by a column (:) like so:
USERNAME:HASH. This was later changed to use base64 encoding of the hash
instead. During this code change, the username prefix (USERNAME:) was
forgotten about and the API would therefore malfunction.

Addition to that, getUser() now fallbacks to the default XenForo visitor
if the API key isn't used.

This fixes #46 and #47